[Gold IV] Title: Symmetry of Stars, Time: 216 ms, Memory: 197476 KB -BaekjoonHub

+++ 백준/Gold/30453. Symmetry of Stars/Symmetry of Stars.cc
@@ -13,7 +13,7 @@ inline ll hsh(pi p)
 	return (((ll)(p.first + 1234567890) * 1234567890 + p.second + 1234567890) % 99999989 + 99999989) % 99999989;
 }
 
-unsigned cnt[100000000];
+short cnt[100000000];
 
 void solve()
 {
